pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["sawaca96 <sawaca96@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = "^0.26.0"
confluent-kafka = "^2.3.0"
redis = "^5.0.1"
faust-streaming = "^0.10.22"
websockets = "^12.0"
orjson = "^3.9.12"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"
black = "23.12.1"
ruff = "0.1.14"
mypy = "1.8.0"
click = "^8.1.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
disallow_any_generics = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
check_untyped_defs = true
strict_optional = false
warn_redundant_casts = true
warn_unused_ignores = true
allow_redefinition = true
show_column_numbers = true
pretty = true
plugins = ["pydantic.mypy"]
